size_t: solr function helpers

+++ src/php7/php_solr.h
@@ -813,7 +813,7 @@ PHP_SOLR_API void solr_document_get_field_names(INTERNAL_FUNCTION_PARAMETERS);
 /* {{{ Utility functions */
 PHP_SOLR_API long solr_get_json_last_error(void);
 PHP_SOLR_API solr_char_t *solr_get_json_error_msg(solr_json_error_codes_t error_code);
-PHP_SOLR_API int solr_json_to_php_native(solr_string_t *buffer, const solr_char_t *json_string, int json_string_length);
+PHP_SOLR_API int solr_json_to_php_native(solr_string_t *buffer, const solr_char_t *json_string, size_t json_string_length);
 PHP_SOLR_API int solr_is_supported_response_writer(const solr_char_t * response_writer, size_t length);
 PHP_SOLR_API int solr_hashtable_get_new_index(HashTable *ht);
 #if PHP_VERSION_ID < 80000
@@ -835,11 +835,11 @@ PHP_SOLR_API void solr_response_get_response_impl(INTERNAL_FUNCTION_PARAMETERS, 
 /* }}} */
 
 /* {{{ SolrFunction Helpers (solrfunc to avoid confusion with solr_function) */
-PHP_SOLR_API int  solr_solrfunc_update_string(zval *obj, solr_char_t *key, int key_len, solr_char_t *value, int value_len);
-PHP_SOLR_API int  solr_solrfunc_fetch_string(zval *obj, solr_char_t *key, int key_len, solr_string_t **string);
-PHP_SOLR_API int  solr_solrfunc_return_string(zval *obj, solr_char_t *key, int key_len, zval **return_value);
+PHP_SOLR_API int  solr_solrfunc_update_string(zval *obj, solr_char_t *key, size_t key_len, solr_char_t *value, size_t value_len);
+PHP_SOLR_API int  solr_solrfunc_fetch_string(zval *obj, solr_char_t *key, size_t key_len, solr_string_t **string);
+PHP_SOLR_API int  solr_solrfunc_return_string(zval *obj, solr_char_t *key, size_t key_len, zval **return_value);
 PHP_SOLR_API void solr_solrfunc_to_string(solr_function_t *function, solr_string_t **dest);
-PHP_SOLR_API int solr_solrfunc_display_string(zval *obj, solr_char_t *key, int key_len, zval **return_value);
+PHP_SOLR_API int solr_solrfunc_display_string(zval *obj, solr_char_t *key, size_t key_len, zval **return_value);
 
 #if PHP_VERSION_ID < 80000
 zend_object *solr_collapse_function_object_handler_clone(zval *object);
